extract tosu state into context provider
